All calc methods returns double values now. But test still checks them as float. Because rest assured can not work with double values.

+++ src/main/java/com/upwork/dto/Result.java
@@ -6,17 +6,17 @@ import java.io.Serializable;
  * Created by vbabin on 11.11.2016.
  */
 public class Result implements Serializable {
-    private float result;
+    private double result;
 
-    public Result(float result) {
+    public Result(double result) {
         this.result = result;
     }
 
-    public float getResult() {
+    public double getResult() {
         return result;
     }
 
-    public void setResult(float result) {
+    public void setResult(double result) {
         this.result = result;
     }
 }
